Study day game 6.5/7
I have further improved the structure of my classes. Now I have:

A general class describing cards.
A scene with a card object created from nodes,
A description of the cards physics,
A database containing a dictionary of all cards classes and the process of their creation
A separate folder with subclasses for each cardtype