build week 2, day 3.1
I finished writing  basic functionality of the collection and described the deck template. Now the deck is modified, can store universal data for all scenes, and save itself outside the game and scene.

Next, I can unload the deck to work with it during the match and develop the main mechanics of the competition