build week 3, day 1
Found the root problem in my project due to the difference in the interpretation of inputs if the control node and 2D node are in the same scene. Fixed it and, with considerable help from Cloud, added the deck, hand, and cards in hand transfer systems. Started writing the card draw. But I will spend tomorrow on reworking and fixing Cloud's system